// src/app/models/task/taskStatus.ts
export enum TaskStatus {
  HIDE = 'HIDE',
  ON_HOLD = 'ON_HOLD',

  PUBLISHED = 'PUBLISHED',
  PERFORMER_SELECTED = 'PERFORMER_SELECTED',
  NOT_PAYED = 'NOT_PAYED',
  PROCESSING = 'PROCESSING',
  REQUEST_TO_FINISH = 'REQUEST_TO_FINISH',
  FINISHED = 'FINISHED',
  CANCELED_BY_PARTNER = 'CANCELED_BY_PARTNER',
  CANCELED_BY_PERFORMER = 'CANCELED_BY_PERFORMER',

  FINISHED_IN_FAVOR_PERFORMER = 'FINISHED_IN_FAVOR_PERFORMER',
  FINISHED_IN_FAVOR_CUSTOMER = 'FINISHED_IN_FAVOR_CUSTOMER'
}

export const TaskStatusesWithAssignedPerformer = [
  TaskStatus.PROCESSING,
  TaskStatus.REQUEST_TO_FINISH,
  TaskStatus.FINISHED,
  TaskStatus.CANCELED_BY_PARTNER,
  TaskStatus.CANCELED_BY_PERFORMER,
  TaskStatus.FINISHED_IN_FAVOR_PERFORMER,
  TaskStatus.FINISHED_IN_FAVOR_CUSTOMER
];