// src/app/models/task/taskRequestStatus.ts
export enum TaskRequestStatus {
  NOT_ENOUGH_MONEY = 'NOT_ENOUGH_MONEY',
  SUBMITTED = 'SUBMITTED',
  APPROVED_BY_CUSTOMER = 'APPROVED_BY_CUSTOMER',
  REJECTED_BY_CUSTOMER = 'REJECTED_BY_CUSTOMER',
  REJECTED_BY_PERFORMER = 'REJECTED_BY_PERFORMER',

  ACCEPTED_BY_PERFORMER = 'ACCEPTED_BY_PERFORMER'
}